Fixed print bug, enhanced UI, and added debug mode - Fixed: When in print mode, selecting from checkboxes only (vs. selecting entire item in list) would not send QRs to printer - Fixed: Checkboxes not showing up on smaller device screens - Tweaked style of main entry list items to better utilize available space   * Moved icon to other side, put title above image, and increased description size - Debug Mode: Set DEBUG to true in MainActivity to create and auto-populate database (or click debug icon)

+++ app/src/main/java/com/github/scanme/database/QRRepository.java
@@ -76,8 +76,10 @@ public class QRRepository {
         protected Void doInBackground(final QR... params) {
             QR qr = params[0];
             asyncDao.delete(qr);
-            deleteFile(qr.getImagePath());
-            deleteFile(qr.getQrPath());
+            if (!qr.getImagePath().startsWith("test")) {
+                deleteFile(qr.getImagePath());
+                deleteFile(qr.getQrPath());
+            }
             return null;
         }
     }
